requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.5=py38h080aedc_0
brotli=1.0.9=ha925a31_2
brotli-python=1.0.9=py38hd77b12b_7
ca-certificates=2023.7.22=h56e8100_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py38h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=py38haa95532_0
contourpy=1.0.5=py38h59b6b97_0
cryptography=41.0.3=py38h89fc84f_0
cudatoolkit=11.1.1=heb2d755_10
cycler=0.12.1=pyhd8ed1ab_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.9.0=py38haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
icu=73.1=h6c2663c_0
idna=3.4=py38haa95532_0
importlib_resources=6.1.1=pyhd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46319
ipython=8.12.2=py38haa95532_0
jedi=0.18.1=py38haa95532_1
jinja2=3.1.2=py38haa95532_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=py38hd77b12b_0
krb5=1.20.1=h5b6d351_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=5_hd5c7e75_netlib
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
liblapack=3.9.0=5_hd5c7e75_netlib
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.7.2=py38haa95532_0
matplotlib-base=3.7.2=py38h4ed8f06_0
matplotlib-inline=0.1.6=py38haa95532_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.8=py38h2bbff1b_0
mkl_random=1.2.4=py38h59b6b97_0
mpmath=1.3.0=py38haa95532_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
networkx=3.1=py38haa95532_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
numexpr=2.8.4=py38h7b80656_1
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.12=h2bbff1b_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.0.3=py38h4ed8f06_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.0.1=py38h045eedc_0
pip=23.3=py38haa95532_0
ply=3.11=py_1
prompt-toolkit=3.0.36=py38haa95532_0
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py38haa95532_1
pyopenssl=23.2.0=py38haa95532_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.10=py38hd77b12b_0
pyqt5-sip=12.13.0=py38h2bbff1b_0
pysocks=1.7.1=py38haa95532_0
python=3.8.18=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.0.1=cpu_py38hb0bdfb8_0
pytz=2023.3.post1=py38haa95532_0
qt-main=5.15.2=h19c9488_10
requests=2.31.0=py38haa95532_0
scikit-learn=0.22.1=py38h7208079_1
scipy=1.8.1=py38h4990095_2
setuptools=68.0.0=py38haa95532_0
sip=6.7.12=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py38haa95532_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.12=h2bbff1b_0
tomli=2.0.1=pyhd8ed1ab_0
torchvision=0.15.2=cpu_py38h7187fe4_0
tornado=6.2=py38h294d835_0
traitlets=5.7.1=py38haa95532_0
typing-extensions=4.7.1=py38haa95532_0
typing_extensions=4.7.1=py38haa95532_0
urllib3=1.26.18=py38haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.41.2=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
xz=5.4.2=h8cc25b3_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
